pyproject.toml
[project]
name = "whatsapp"
version = "0.1.0"
description = "A real-time voice bot that integrates with WhatsApp Business API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[google,silero, webrtc]>=0.0.83",
    "openai>=1.52.0,<2.0.0",
    "faiss-cpu>=1.8.0,<2.0.0",
    "pypdf>=5.1.0,<6.0.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
    "uvicorn>=0.32.0,<1.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "fastapi>=0.115.6,<0.117.0"
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]